requirements.txt
streamlit
faiss-cpu
sentence-transformers
langchain-text-splitters
rank-bm25
numpy
beautifulsoup4
lxml
pdfplumber
tiktoken